Trying again to remove chip when no citations from one of the sources
