+++ src/components/content/chat/chatbubbles/citations/ToggleCitations.tsx
@@ -1,13 +1,34 @@
 import { useState } from "react"
 import { Chips } from "@navikt/ds-react";
+import { Message } from "../../../../../types/Message.ts"
 
 interface ToggleCitationsProps {
   onToggle: (selected: string[]) => void
-  options: string[]
+  message: Message
 }
 
-const ToggleCitations = ({ onToggle, options }: ToggleCitationsProps ) => {
-  const [selected, setSelected] = useState<string[]>(options)
+const citationOptions = [
+  "Sitater fra Nav.no",
+  "Sitater fra Kunnskapsbasen",
+];
+
+const ToggleCitations = ({ onToggle, message }: ToggleCitationsProps ) => {
+  const hasNavnoCitations = message.citations.some(
+    (citation) => message.context[citation.sourceId].source === "navno"
+  )
+  const hasKunnskapsbasenCitations = message.citations.some(
+    (citation) => message.context[citation.sourceId].source === "nks"
+  )
+
+  const filteredOptions = citationOptions.filter((option) => {
+    if (option === "Sitater fra Nav.no" && !hasNavnoCitations) {
+      return false;
+    }
+    return !(option === "Sitater fra Kunnskapsbasen" && !hasKunnskapsbasenCitations);
+
+  })
+
+  const [selected, setSelected] = useState<string[]>(filteredOptions)
 
   const handleToggle = (option: string) => {
     const newSelected = selected.includes(option)
@@ -20,7 +41,7 @@ const ToggleCitations = ({ onToggle, options }: ToggleCitationsProps ) => {
   return (
     <div className='mb-3'>
       <Chips size='small'>
-        {options.map((option) => (
+        {filteredOptions.map((option) => (
           <Chips.Toggle
             key={option}
             selected={selected.includes(option)}
